feat: 기존의 리포지 토리를 서비스와 리포지토리로 나눴습니다.
기능을 구현하다보니 자료구조 List와 회원가입 하는 로직을 분리 해야겠다는 생각이들어 둘을 분리했습니다.
